+++ tests/test_marshalling.py
@@ -6,11 +6,7 @@ import pytest
 from flask_restplus import (
     marshal, marshal_with, marshal_with_field, fields, Api, Resource
 )
-
-try:
-    from collections import OrderedDict
-except ImportError:
-    from ordereddict import OrderedDict
+from flask_restplus._compat import OrderedDict
 
 
 # Add a dummy Resource to verify that the app is properly set.