+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Changelog
 Current
 -------
 
+- Performance improvements:
+    - Make use of Cython when available
 - Add vendor extensions support (:issue:`97`)
 - :class:`~reqparse.RequestParser` arguments now support the ``split`` action
 - Ensure default boolean value as `False` works with :class:`~reqparse.RequestParser` (:issue:`199`)